表字段维护

+++ src/main/java/com/hxsd/controller/QuestionController.java
@@ -21,7 +21,7 @@ public class QuestionController {
     private QuestionService questionService;
 
     @GetMapping("/question/{id}")
-    public String profile(@PathVariable(name = "id") Integer id,
+    public String profile(@PathVariable(name = "id") Long id,
                           Model model,
                           HttpServletRequest request,
                           @RequestParam(name = "page",defaultValue = "1")Integer page,
